syntax adjustment for 1.8.7
Forgot that 1.8.7 doesn't have require_relative.  Changing the specs
to use load, and to target the final script (`git-issue`, in the
base directory), but load needs a file extension: so now there's a
symlink in the spec directory `target.rb -> ../git-issue`

also need to swap new hash syntax for old, and apparently 1.8.7
freaks out if you use "" with a \# inside that doesn't have the
curly-braces next to it. Also threw up on strings that start with
an escaped newline.